pyproject.toml
[project]
name = "morsels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "atproto>=0.0.65",
    "authlib>=1.6.9",
    "flask>=3.1.3",
    "gunicorn>=25.3.0",
    "pygments>=2.20.0",
    "regex>=2026.3.32",
    "requests-hardened>=1.2.0",
]
